requirements.txt
azure-functions
joblib
numpy